Разделение бизнес-логики и контроллеров реквестов
